+++ testnet/stacks-node/src/tests/signer/v0.rs
@@ -49,7 +49,6 @@ use stacks::chainstate::stacks::boot::MINERS_NAME;
 use stacks::chainstate::stacks::db::{StacksBlockHeaderTypes, StacksChainState, StacksHeaderInfo};
 use stacks::chainstate::stacks::miner::{
     TransactionEvent, TransactionSuccessEvent, TEST_EXCLUDE_REPLAY_TXS,
-    TEST_MINE_ALLOWED_REPLAY_TXS,
 };
 use stacks::chainstate::stacks::{
     StacksTransaction, TenureChangeCause, TenureChangePayload, TransactionPayload,
@@ -17350,6 +17349,13 @@ fn bitcoin_reorg_extended_tenure() {
     info!("Bitcoin fork triggered"; "ch" => %before_fork, "btc_height" => burn_block_height);
     info!("Chain info before fork: {:?}", get_chain_info(&conf_1));
 
+    // Make sure signers don't perform block broadcast for the next bits:
+    //  we want to ensure that the *miner* is the one broadcast blocks,
+    //  because when we stall p2p broadcast, we don't want to accidentally
+    //  stall the miner in the situation where they produce block A, signers broadcast it,
+    //  we initiate the stall, and then the miner attempts to broadcast A.
+    stacks_signer::v0::tests::TEST_SKIP_BLOCK_BROADCAST.set(true);
+
     let mut after_fork = get_chain_info(&conf_1).pox_consensus;
     wait_for(60, || {
         after_fork = get_chain_info(&conf_1).pox_consensus;
@@ -17372,7 +17378,8 @@ fn bitcoin_reorg_extended_tenure() {
     //  so that we can ensure all the signers approve the proposal
     //  before it gets accepted by stacks-nodes
     TEST_P2P_BROADCAST_STALL.set(true);
-    stacks_signer::v0::tests::TEST_SKIP_BLOCK_BROADCAST.set(true);
+
+    info!("Stalled broadcast, submitting a contract call!");
 
     // the signer signature hash is the same as the block header hash.
     // we use the latest_signer_sighash to make sure we're getting block responses for the
@@ -17393,11 +17400,18 @@ fn bitcoin_reorg_extended_tenure() {
     let rc = miners.signer_test.get_current_reward_cycle();
     let slot_ids = miners.signer_test.get_signer_indices(rc);
     let mut block_responses: Vec<_> = vec![];
+
     wait_for(60, || {
         block_responses = slot_ids
             .iter()
             .filter_map(|slot_id| {
                 let latest_br = miners.signer_test.get_latest_block_response(slot_id.0);
+                info!(
+                    "[{}] Checking response for {}. accepted = {}",
+                    slot_id.0,
+                    latest_br.get_signer_signature_hash(),
+                    latest_br.as_block_accepted().is_some()
+                );
                 if latest_br.get_signer_signature_hash() != latest_signer_sighash {
                     Some(latest_br)
                 } else {